DreamHack/Level4/OTT/external/requirements.txt
flask==2.3.3
python-docx==0.8.11
requests==2.31.0 